+++ app/Plugin.php
@@ -3,11 +3,15 @@
 namespace App;
 
 use App\File\Directory;
+use App\Services\AccessPointsService;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
+
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
@@ -36,19 +40,19 @@ class Plugin extends Model
         'licence',
         'title',
     ];
-    
+
     private static function pluginDirectory() {
         $pluginDirectory = config('app.plugin_directory');
         return base_path($pluginDirectory);
     }
-    
+
     public static function getPluginPath(string $path = ''):string {
         if($path === ''){
             return self::pluginDirectory();
         }
         return self::pluginDirectory() . Str::start($path, '/');
     }
-    
+
     public static function isInstalled($name): bool {
         return self::whereNotNull('installed_at')->where('name', $name)->exists();
     }
@@ -60,7 +64,7 @@ class Plugin extends Model
     public function slugName(): string {
         return Str::slug($this->name);
     }
-    
+
     public function getPath(string $path = ''): string {
         $pluginPath = $this->name;
         if($path !== ''){
@@ -92,7 +96,7 @@ class Plugin extends Model
 
         return json_decode(json_encode($xmlObject), true);
     }
-    
+
     public function getInfo(){
         return self::getPluginInfo($this->getPath());
     }
@@ -107,7 +111,7 @@ class Plugin extends Model
                         $metadata[$field] = [];
                         continue;
                     }
-                    
+
                     $authors = $info[$field]['author'];
                     $metadata[$field] = is_array($authors) ? $authors : [$authors];
                 } else {
@@ -134,6 +138,110 @@ class Plugin extends Model
             }
         }
         return $changes;
+    }
+
+    public function getAccessPoints(): array {
+        $info = self::getInfo();
+        $accesspoints = [];
+        $addedNames = [];
+        $addedPaths = [];
+        if($info !== false) {
+            if(array_key_exists('accesspoints', $info)) {
+                foreach($info['accesspoints'] as $accesspoint) {
+                    $name = $this->name . '-' . $accesspoint['id'];
+                    $label = $accesspoint['label'];
+                    $path = Str::finish(Str::start($accesspoint['path'], '/'), '/');
+                    // $path = '/' . $this->slugName() . Str::finish(Str::start($accesspoint['url'], '/'), '/');
+                    if(array_key_exists($name, $addedNames)) {
+                        throw new \Exception("An accesspoint with the name ($name) already exists");
+                    }
+                    if(array_key_exists($path, $addedPaths)) {
+                        throw new \Exception("An accesspoint with the path ($path) already exists");
+                    }
+
+                    $addedNames[$name] = true;
+                    $addedPaths[$path] = true;
+
+                    $accesspoints[$name] = [
+                        'label' => $label,
+                        'path' => $path,
+                    ];
+                }
+            }
+        }
+        return $accesspoints;
+    }
+
+    private function getScopeCacheKey(): string {
+        return 'plugin_scopes_' . $this->id;
+    }
+
+    public function getScopes(): array {
+        return Cache::rememberForever($this->getScopeCacheKey(), function() {
+            $info = self::getInfo();
+            $scopes = [];
+            if($info !== false) {
+                if(array_key_exists('scopes', $info)) {
+                    foreach($info['scopes'] as $scope) {
+                        $attributes = $scope['@attributes'];
+                        if(!array_key_exists('src', $attributes)) {
+                            Log::error('<scope> attribute \'src\' is required');
+                            continue;
+                        }
+                        if(!array_key_exists('on', $attributes)) {
+                            Log::error('<scope> attribute \'on\' is required');
+                            continue;
+                        }
+
+                        $src = $attributes['src'];
+                        $on = $attributes['on'];
+
+                        $srcDir = $this->getPath("Scopes");
+                        if(!file_exists($srcDir) || !is_dir($srcDir)) {
+                            Log::error('Missing \'Scopes\' directory');
+                            continue;
+                        }
+                        $srcPath = $srcDir . DIRECTORY_SEPARATOR . $src;
+                        if(!file_exists($srcPath)) {
+                            Log::error("Missing file '$src'");
+                            continue;
+                        }
+                        if(!class_exists($on)) {
+                            Log::error("Class '{$on}' does not exist!");
+                            continue;
+                        }
+                        $className = Str::replaceEnd('.php', '', $src);
+                        $namespacedSrc = "App\\Plugins\\$this->name\\Scopes\\$className";
+
+                        if(!array_key_exists($on, $scopes)) {
+                            $scopes[$on] = [];
+                        }
+
+                        $scopes[$on][] = $namespacedSrc;
+                    }
+                }
+            }
+            return $scopes;
+        });
+    }
+
+    /**
+     * Get all scopes defined in Plugins for a given model class (e.g. App\Entity).
+     */
+    public static function getScopesFor(string $modelClass) {
+        $scopes = [];
+
+        $installedPlugins = Plugin::getInstalled();
+        foreach($installedPlugins as $plugin) {
+            $pluginScopes = $plugin->getScopes();
+            if(array_key_exists($modelClass, $pluginScopes)) {
+                foreach($pluginScopes[$modelClass] as $scope) {
+                    $scopes[] = $scope;
+                }
+            }
+        }
+
+        return $scopes;
     }
 
     public function getRegisteredAttributes(): array {
@@ -244,11 +352,20 @@ class Plugin extends Model
         }
     }
 
+    public function clearCache(): void {
+        Cache::forget($this->getScopeCacheKey());
+        
+        // TODO: Models are meant for the data layer only, 
+        // we should restructure the code into service classes (e.g. PluginMigrationService) 
+        app(AccessPointsService::class)->clearCache();
+    }
+
     public function handleInstallation(bool $isUpdate = false): void {
         $this->runMigrations();
         $this->publishScript();
         $this->addPermissions();
         $this->installPresetsFromFile();
+        $this->clearCache();
 
         if(!$isUpdate) {
             $this->installed_at = Carbon::now();
@@ -269,7 +386,7 @@ class Plugin extends Model
 
     public function handleUninstall(): void {
         $this->removeScript();
-
+        $this->clearCache();
         $this->installed_at = null;
         $this->save();
     }
@@ -372,7 +489,7 @@ class Plugin extends Model
 
             self::getDirectory()->store(
                 $this->publicName(false),
-                $scriptPath
+                $filehandle
             );
             fclose($filehandle);
         } else {