+++ tests/Feature/ApiUserTest.php
@@ -183,7 +183,8 @@ class ApiUserTest extends TestCase
                 'email' => 'test@test.com',
                 'name' => 'Test User',
                 'nickname' => 'tuser',
-                'password' => 'test1234' // at least 6 characters
+                'password' => 'test1234', // at least 6 characters
+                'password_confirm' => 'test1234',
             ]);
 
         $user = User::latest()->first();
@@ -359,6 +360,7 @@ class ApiUserTest extends TestCase
             'deleted_at' => null,
             'avatar' => null,
             'login_attempts' => null,
+            'accesspoints' => null,
             'metadata' => [
                 'phonenumber' => '+43 123 1234',
                 'orcid' => '0000-0002-1694-233X',